v0.0.5 版本: 1. 修复 bug; 2. 添加部分注释; 3. 默认去除 mongodb 与 redis 依赖;

+++ agtms-parent/agtms-core/src/main/java/net/saisimon/agtms/core/enums/Classes.java
@@ -1,8 +1,14 @@
 package net.saisimon.agtms.core.enums;
 
+/**
+ * 模版列属性类型枚举
+ * 
+ * @author saisimon
+ *
+ */
 public enum Classes {
 	
-	INTEGER("int"), DATE("date"), STRING("string"), DOUBLE("double");
+	LONG("long"), DATE("date"), STRING("string"), DOUBLE("double");
 	
 	private String name;
